DATAJPA-1649 Ensure repository declares correct ID type.

+++ src/test/java/org/springframework/data/jpa/repository/query/JpaCountQueryCreatorIntegrationTests.java
@@ -70,7 +70,7 @@ public class JpaCountQueryCreatorIntegrationTests {
 		assertThat(HibernateUtils.getHibernateQuery(query)).startsWith("select distinct count(distinct");
 	}
 
-	interface SomeRepository extends Repository<User, Long> {
+	interface SomeRepository extends Repository<User, Integer> {
 		void findDistinctByRolesIn(List<Role> roles);
 	}
 }
